# Use callback function instead of retun values in common-prefix search