re: #75 

before
```
$ go test --bench .
PASS
BenchmarkAnalyzeNormal-4         10000        222285 ns/op
BenchmarkAnalyzeSearch-4          5000        290241 ns/op
BenchmarkAnalyzeExtended-4        5000        295677 ns/op
ok      github.com/ikawaha/kagome/tokenizer    21.251s
```

after
```
go test --bench .
PASS
BenchmarkAnalyzeNormal-4         10000        158133 ns/op
BenchmarkAnalyzeSearch-4         10000        227858 ns/op
BenchmarkAnalyzeExtended-4       10000        235153 ns/op
ok      github.com/ikawaha/kagome/tokenizer    22.270s
```

alloc_space
```
$ go tool pprof --alloc_space tokenizer.test fix_mem.prof
Entering interactive mode (type "help" for commands)
(pprof) top
1020.66MB of 1033.17MB total (98.79%)
Dropped 27 nodes (cum <= 5.17MB)
Showing top 10 nodes out of 44 (cum >= 7.50MB)
      flat  flat%   sum%        cum   cum%
  773.95MB 74.91% 74.91%   775.96MB 75.10%  github.com/ikawaha/kagome/tokenizer.Tokenizer.Analyze
  114.31MB 11.06% 85.97%   114.31MB 11.06%  bytes.makeSlice
   60.99MB  5.90% 91.88%    60.99MB  5.90%  strings.genSplit
   46.50MB  4.50% 96.38%    46.50MB  4.50%  encoding/binary.Read
   10.34MB  1.00% 97.38%    32.84MB  3.18%  github.com/ikawaha/kagome/internal/da.Read
    8.98MB  0.87% 98.25%    69.97MB  6.77%  github.com/ikawaha/kagome/internal/dic.NewContents
    3.31MB  0.32% 98.57%    20.31MB  1.97%  github.com/ikawaha/kagome/internal/dic.LoadConnectionTable
    2.28MB  0.22% 98.79%     9.28MB   0.9%  github.com/ikawaha/kagome/internal/dic.LoadMorphSlice
         0     0% 98.79%   117.31MB 11.35%  bytes.(*Buffer).ReadFrom
         0     0% 98.79%     7.50MB  0.73%  encoding/gob.(*Decoder).Decode
```